+tests

+++ src/main/java/ru/itsjava/repository/UserRepositoryImpl.java
@@ -46,11 +46,11 @@ public class UserRepositoryImpl implements UserRepository {
     public Optional<User> findUserByEmail(Email email) {
         Query queryUser = entityManager.createQuery("select id from users where email = :email");
         queryUser.setParameter("email", email);
-        List <Long> resultList = queryUser.getResultList();
+        List<Long> resultList = queryUser.getResultList();
         if (resultList.isEmpty()) {
             return Optional.empty();
         } else {
-            Long foundUserId = (Long) resultList.get(0);
+            Long foundUserId = resultList.get(0);
             User foundUser = entityManager.find(User.class, foundUserId);
             return Optional.ofNullable(foundUser);
         }
